Fix devise initializer so the flashes would come up during the signup/login processes
A while ago when I was replacing turbolinks with turbo-rails, I needed to configure devise initializer as well to work with it, I wrote the code and did not test it properly. However, only when writing cypress tests I realised no flashes were coming up when signing up and logging in. With this additional line flashes are coming up now